+++ src/gui/CentralWidget.cpp
@@ -90,7 +90,7 @@ CentralWidget::CentralWidget(QWidget *parent) : QWidget(parent)
 
         l->addWidget(m, 1, 0, 1, 6, Qt::AlignLeft);
 
-        connect(c, &Connector::updateBlockReq, this, [this](int ii) {
+        connect(c, &Connector::updateBlockReq, this, [this](long unsigned int ii) {
             m->updateBlock(ii);
         });
     });
